Inlining interfaces
