--- src/Rn.DnsUpdater.Core/Services/Interfaces/IConfigService.cs
@@ -1,13 +1,0 @@
-using Rn.DnsUpdater.Core.Config;
-
-namespace Rn.DnsUpdater.Core.Services.Interfaces;
-
-public interface IConfigService
-{
-  DnsUpdaterConfig CoreConfig { get; }
-  DnsEntriesConfig DnsEntriesConfig { get; }
-
-  List<DnsUpdaterEntry> GetEntriesNeedingUpdate();
-  List<DnsUpdaterEntry> GetEnabledEntries();
-  void SaveConfigState();
-}
