--- src/Rn.DnsUpdater.Core/Services/Interfaces/IDnsUpdaterService.cs
@@ -1,8 +1,0 @@
-﻿using Rn.DnsUpdater.Core.Config;
-
-namespace Rn.DnsUpdater.Core.Services.Interfaces;
-
-public interface IDnsUpdaterService
-{
-  Task UpdateEntryAsync(DnsUpdaterEntry entry, CancellationToken stoppingToken);
-}
