--- src/Rn.DnsUpdater.Core/Services/Interfaces/IHostIpAddressService.cs
@@ -1,6 +1,0 @@
-﻿namespace Rn.DnsUpdater.Core.Services.Interfaces;
-
-public interface IHostIpAddressService
-{
-  Task<bool> HostAddressChangedAsync(CancellationToken stoppingToken);
-}
